Add ChunkGenerationBatch, replace batches with 'v2' ones, remove batch creation methods in Instance
