+++ src/main/java/net/minestom/server/instance/Chunk.java
@@ -9,8 +9,6 @@ import net.minestom.server.entity.pathfinding.PFColumnarSpace;
 import net.minestom.server.event.player.PlayerChunkLoadEvent;
 import net.minestom.server.event.player.PlayerChunkUnloadEvent;
 import net.minestom.server.instance.batch.BatchOption;
-import net.minestom.server.instance.batch.BlockBatch;
-import net.minestom.server.instance.batch.ChunkBatch;
 import net.minestom.server.instance.block.Block;
 import net.minestom.server.instance.block.BlockManager;
 import net.minestom.server.instance.block.CustomBlock;
@@ -102,7 +100,7 @@ public abstract class Chunk implements Viewable, DataContainer {
      * <p>
      * This is used when the previous block has to be destroyed/replaced, meaning that it clears the previous data and update method.
      * <p>
-     * WARNING: this method is not thread-safe (in order to bring performance improvement with {@link ChunkBatch} and {@link BlockBatch})
+     * WARNING: this method is not thread-safe (in order to bring performance improvement with {@link net.minestom.server.instance.batch.Batch}s)
      * The thread-safe version is {@link InstanceContainer#setSeparateBlocks(int, int, int, short, short, Data)} (or any similar instance methods)
      * Otherwise, you can simply do not forget to have this chunk synchronized when this is called.
      *
@@ -253,8 +251,6 @@ public abstract class Chunk implements Viewable, DataContainer {
 
     /**
      * Resets the chunk, this means clearing all the data making it empty.
-     * <p>
-     * Used for {@link BatchOption#isFullChunk()}.
      */
     public abstract void reset();
