+++ src/main/java/net/minestom/server/instance/Instance.java
@@ -16,8 +16,6 @@ import net.minestom.server.event.handler.EventHandler;
 import net.minestom.server.event.instance.AddEntityToInstanceEvent;
 import net.minestom.server.event.instance.InstanceTickEvent;
 import net.minestom.server.event.instance.RemoveEntityFromInstanceEvent;
-import net.minestom.server.instance.batch.BlockBatch;
-import net.minestom.server.instance.batch.ChunkBatch;
 import net.minestom.server.instance.block.Block;
 import net.minestom.server.instance.block.BlockManager;
 import net.minestom.server.instance.block.CustomBlock;
@@ -211,22 +209,6 @@ public abstract class Instance implements BlockModifier, EventHandler, DataConta
      * @param callback optional callback called when the chunks are done saving
      */
     public abstract void saveChunksToStorage(@Nullable Runnable callback);
-
-    /**
-     * Creates a new {@link BlockBatch} linked to this instance.
-     *
-     * @return a {@link BlockBatch} linked to the instance
-     */
-    public abstract BlockBatch createBlockBatch();
-
-    /**
-     * Creates a new {@link Chunk} batch linked to this instance and the specified chunk.
-     *
-     * @param chunk the chunk to modify
-     * @return a ChunkBatch linked to {@code chunk}
-     * @throws NullPointerException if {@code chunk} is null
-     */
-    public abstract ChunkBatch createChunkBatch(@NotNull Chunk chunk);
 
     /**
      * Gets the instance {@link ChunkGenerator}.
