+++ src/main/java/net/minestom/server/instance/batch/ChunkBatch.java
@@ -5,210 +5,166 @@ import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 import it.unimi.dsi.fastutil.longs.LongList;
 import net.minestom.server.data.Data;
-import net.minestom.server.instance.*;
-import net.minestom.server.instance.block.CustomBlock;
+import net.minestom.server.instance.Chunk;
+import net.minestom.server.instance.InstanceContainer;
 import net.minestom.server.utils.block.CustomBlockUtils;
 import net.minestom.server.utils.callback.OptionalCallback;
 import net.minestom.server.utils.chunk.ChunkCallback;
 import net.minestom.server.utils.chunk.ChunkUtils;
-import net.minestom.server.utils.validate.Check;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * Used when all the blocks you want to place can be contained within only one {@link Chunk},
- * use a {@link BlockBatch} instead otherwise.
- * Can be created using {@link Instance#createChunkBatch(Chunk)}, and executed with {@link #flush(ChunkCallback)}.
+ * A Batch used when all of the block changed are contained inside a single chunk.
+ * If more than one chunk is needed, use an {@link AbsoluteBlockBatch} instead.
  * <p>
- * Uses chunk coordinate (0-15) instead of world's.
+ * The batch can be placed in any chunk in any instance, however it will always remain
+ * aligned to a chunk border. If completely translatable block changes are needed, use a
+ * {@link RelativeBlockBatch} instead.
+ * <p>
+ * Coordinates are relative to the chunk (0-15) instead of world coordinates.
  *
- * @see InstanceBatch
+ * @see Batch
  */
-public class ChunkBatch implements InstanceBatch {
-
-    private final InstanceContainer instance;
-    private final Chunk chunk;
-    private final BatchOption batchOption;
-
-    private final boolean generationBatch;
+public class ChunkBatch implements Batch<ChunkCallback> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ChunkBatch.class);
 
     // Need to be synchronized manually
     // Format: blockIndex/blockStateId/customBlockId (32/16/16 bits)
-    private LongList blocks;
+    private final LongList blocks;
 
     // Need to be synchronized manually
     // block index - data
-    private Int2ObjectMap<Data> blockDataMap;
+    private final Int2ObjectMap<Data> blockDataMap;
 
-    public ChunkBatch(@NotNull InstanceContainer instance, @NotNull Chunk chunk,
-                      @NotNull BatchOption batchOption,
-                      boolean generationBatch) {
-        this.instance = instance;
-        this.chunk = chunk;
-        this.batchOption = batchOption;
-        this.generationBatch = generationBatch;
-
-        if (!generationBatch) {
-            this.blocks = new LongArrayList();
-            this.blockDataMap = new Int2ObjectOpenHashMap<>();
-        }
+    public ChunkBatch() {
+        this(new LongArrayList(), new Int2ObjectOpenHashMap<>());
     }
 
-    public ChunkBatch(@NotNull InstanceContainer instance, @NotNull Chunk chunk,
-                      boolean generationBatch) {
-        this(instance, chunk, new BatchOption(), generationBatch);
-    }
-
-    @Override
-    public void setBlockStateId(int x, int y, int z, short blockStateId, @Nullable Data data) {
-        addBlockData((byte) x, y, (byte) z, blockStateId, (short) 0, data);
-    }
-
-    @Override
-    public void setCustomBlock(int x, int y, int z, short customBlockId, @Nullable Data data) {
-        final CustomBlock customBlock = BLOCK_MANAGER.getCustomBlock(customBlockId);
-        Check.notNull(customBlock, "The custom block with the id " + customBlockId + " does not exist!");
-        addBlockData((byte) x, y, (byte) z, customBlock.getDefaultBlockStateId(), customBlockId, data);
+    protected ChunkBatch(LongList blocks, Int2ObjectMap<Data> blockDataMap) {
+        this.blocks = blocks;
+        this.blockDataMap = blockDataMap;
     }
 
     @Override
     public void setSeparateBlocks(int x, int y, int z, short blockStateId, short customBlockId, @Nullable Data data) {
-        addBlockData((byte) x, y, (byte) z, blockStateId, customBlockId, data);
-    }
+        // Cache the entry to be placed later during flush
+        final int index = ChunkUtils.getBlockIndex(x, y, z);
+        long value = index;
+        value = (value << 16) | blockStateId;
+        value = (value << 16) | customBlockId;
 
-    private void addBlockData(byte x, int y, byte z, short blockStateId, short customBlockId, @Nullable Data data) {
-        if (isGenerationBatch()) {
-            // Directly place the block
-            chunk.UNSAFE_setBlock(x, y, z, blockStateId, customBlockId, data, CustomBlockUtils.hasUpdate(customBlockId));
-        } else {
-            // Cache the entry to be placed later during flush
+        synchronized (blocks) {
+            this.blocks.add(value);
+        }
 
-            final int index = ChunkUtils.getBlockIndex(x, y, z);
-
-            if (data != null) {
-                synchronized (blockDataMap) {
-                    this.blockDataMap.put(index, data);
-                }
-            }
-
-            long value = index;
-            value = (value << 16) | blockStateId;
-            value = (value << 16) | customBlockId;
-
-            synchronized (blocks) {
-                this.blocks.add(value);
+        if (data != null) {
+            synchronized (blockDataMap) {
+                this.blockDataMap.put(index, data);
             }
         }
     }
 
-    /**
-     * Gets if this chunk batch is part of a chunk generation.
-     * <p>
-     * Being a generation batch mean that blocks set are not being stored
-     * but are immediately placed on the chunks. Using less memory
-     * and CPU cycles.
-     *
-     * @return true if this batch is part of a chunk generation
-     */
-    public boolean isGenerationBatch() {
-        return generationBatch;
-    }
-
-    /**
-     * Called to fill the chunk batch.
-     *
-     * @param chunkGenerator the chunk generator
-     * @param callback       the optional callback executed once the batch is done
-     */
-    public void flushChunkGenerator(@NotNull ChunkGenerator chunkGenerator, @Nullable ChunkCallback callback) {
-        BLOCK_BATCH_POOL.execute(() -> {
-            synchronized (chunk) {
-                final List<ChunkPopulator> populators = chunkGenerator.getPopulators();
-                final boolean hasPopulator = populators != null && !populators.isEmpty();
-
-                if (batchOption.isFullChunk()) {
-                    this.chunk.reset();
-                }
-
-                chunkGenerator.generateChunkData(this, chunk.getChunkX(), chunk.getChunkZ());
-
-                if (hasPopulator) {
-                    for (ChunkPopulator chunkPopulator : populators) {
-                        chunkPopulator.populateChunk(this, chunk);
-                    }
-                }
-
-                updateChunk(callback, true);
-            }
-        });
-    }
-
-    /**
-     * Executes the batch in the dedicated pool and run the callback during the next instance update when blocks are placed
-     * (which means that the callback can be called 50ms after, but in a determinable thread).
-     *
-     * @param callback the callback to execute once the blocks are placed
-     */
-    public void flush(@Nullable ChunkCallback callback) {
-        Check.stateCondition(generationBatch, "#flush is not support for generation batch.");
-        BLOCK_BATCH_POOL.execute(() -> singleThreadFlush(callback, true));
-    }
-
-    /**
-     * Executes the batch in the dedicated pool and run the callback once blocks are placed (in the block batch pool).
-     * <p>
-     * So the callback is executed in an unexpected thread, but you are sure that it will be called immediately.
-     *
-     * @param callback the callback to execute once the blocks are placed
-     */
-    public void unsafeFlush(@Nullable ChunkCallback callback) {
-        Check.stateCondition(generationBatch, "#unsafeFlush is not support for generation batch.");
-        BLOCK_BATCH_POOL.execute(() -> singleThreadFlush(callback, false));
-    }
-
-    /**
-     * Resets the chunk batch by removing all the entries.
-     */
-    public void clearData() {
-        Check.stateCondition(generationBatch, "#clearData is not support for generation batch.");
+    @Override
+    public void clear() {
         synchronized (blocks) {
             this.blocks.clear();
         }
     }
 
     /**
-     * Executes the batch in the current thread.
+     * Apply this batch to chunk (0, 0).
      *
-     * @param callback     the callback to execute once the blocks are placed
-     * @param safeCallback true to run the callback in the instance update thread, otherwise run in the current one
+     * @param instance The instance in which the batch should be applied
+     * @param callback The callback to be executed when the batch is applied
      */
-    private void singleThreadFlush(@Nullable ChunkCallback callback, boolean safeCallback) {
+    @Override
+    public void apply(@NotNull InstanceContainer instance, @Nullable ChunkCallback callback) {
+        apply(instance, 0, 0, callback);
+    }
+
+    /**
+     * Apply this batch to the given chunk.
+     *
+     * @param instance The instance in which the batch should be applied
+     * @param chunkX The x chunk coordinate of the target chunk
+     * @param chunkZ The z chunk coordinate of the target chunk
+     * @param callback The callback to be executed when the batch is applied.
+     */
+    public void apply(@NotNull InstanceContainer instance, int chunkX, int chunkZ, @Nullable ChunkCallback callback) {
+        final Chunk chunk = instance.getChunk(chunkX, chunkZ);
+        if (chunk == null) {
+            LOGGER.warn("Unable to apply ChunkBatch to unloaded chunk ({}, {}) in {}.", chunkX, chunkZ, instance.getUniqueId());
+            return;
+        }
+        apply(instance, chunk, callback);
+    }
+
+    /**
+     * Apply this batch to the given chunk.
+     *
+     * @param instance The instance in which the batch should be applied
+     * @param chunk The target chunk
+     * @param callback The callback to be executed when the batch is applied
+     */
+    public void apply(@NotNull InstanceContainer instance, @NotNull Chunk chunk, @Nullable ChunkCallback callback) {
+        apply(instance, chunk, callback, true);
+    }
+
+    /**
+     * Apply this batch to the given chunk, and execute the callback
+     * immediately when the blocks have been applied, in an unknown thread.
+     *
+     * @param instance The instance in which the batch should be applied
+     * @param chunk The target chunk
+     * @param callback The callback to be executed when the batch is applied
+     */
+    public void unsafeApply(@NotNull InstanceContainer instance, @NotNull Chunk chunk, @Nullable ChunkCallback callback) {
+        apply(instance, chunk, callback, false);
+    }
+
+    /**
+     * Apply this batch to the given chunk, and execute the callback depending on safeCallback.
+     *
+     * @param instance The instance in which the batch should be applied
+     * @param chunk The target chunk
+     * @param callback The callback to be executed when the batch is applied
+     * @param safeCallback If true, the callback will be executed in the next instance update. Otherwise it will be executed immediately upon completion
+     */
+    protected void apply(@NotNull InstanceContainer instance, @NotNull Chunk chunk, @Nullable ChunkCallback callback, boolean safeCallback) {
+        BLOCK_BATCH_POOL.execute(() -> singleThreadFlush(instance, chunk, callback, safeCallback));
+    }
+
+    /**
+     * Applies this batch in the current thread, executing the callback upon completion.
+     */
+    private void singleThreadFlush(InstanceContainer instance, Chunk chunk, @Nullable ChunkCallback callback, boolean safeCallback) {
         if (blocks.isEmpty()) {
             OptionalCallback.execute(callback, chunk);
             return;
         }
 
-        synchronized (chunk) {
-            if (!chunk.isLoaded())
-                return;
-
-            synchronized (blocks) {
-                for (long block : blocks) {
-                    apply(chunk, block);
-                }
-            }
-
-            updateChunk(callback, safeCallback);
+        if (!chunk.isLoaded()) {
+            LOGGER.warn("Unable to apply ChunkBatch to unloaded chunk ({}, {}) in {}.", chunk.getChunkX(), chunk.getChunkZ(), instance.getUniqueId());
+            return;
         }
+
+        synchronized (blocks) {
+            for (long block : blocks) {
+                apply(chunk, block);
+            }
+        }
+
+        updateChunk(instance, chunk, callback, safeCallback);
     }
 
     /**
-     * Places a block which is encoded in a long.
+     * Applies a single block change given a chunk and a value in the described format.
      *
-     * @param chunk the chunk to place the block on
-     * @param value the block data
+     * @param chunk The chunk to apply the change
+     * @param value block index|state id|custom block id (32|16|16 bits)
      */
     private void apply(@NotNull Chunk chunk, long value) {
         final short customBlockId = (short) (value & 0xFFFF);
@@ -228,24 +184,24 @@ public class ChunkBatch implements InstanceBatch {
                 blockId, customBlockId, data, CustomBlockUtils.hasUpdate(customBlockId));
     }
 
-    private void updateChunk(@Nullable ChunkCallback callback, boolean safeCallback) {
-
+    /**
+     * Updates the given chunk for all of its viewers, and executes the callback.
+     */
+    private void updateChunk(InstanceContainer instance, Chunk chunk, @Nullable ChunkCallback callback, boolean safeCallback) {
         // Refresh chunk for viewers
-        if (batchOption.isFullChunk()) {
-            chunk.sendChunk();
-        } else {
-            chunk.sendChunkUpdate();
-        }
 
-        this.instance.refreshLastBlockChangeTime();
+        // Formerly this had an option to do a Chunk#sendChunkUpdate
+        // however Chunk#sendChunk does the same including a light update
+        chunk.sendChunk();
+
+        instance.refreshLastBlockChangeTime();
 
         if (callback != null) {
             if (safeCallback) {
-                this.instance.scheduleNextTick(inst -> callback.accept(chunk));
+                instance.scheduleNextTick(inst -> callback.accept(chunk));
             } else {
                 callback.accept(chunk);
             }
         }
     }
-
 }
