+++ src/main/java/net/minestom/server/instance/batch/RelativeBlockBatch.java
@@ -1,0 +1,4 @@
+package net.minestom.server.instance.batch;
+
+public class RelativeBlockBatch {
+}
