--- src/main/java/net/minestom/server/instance/batch/v2/RelativeBlockBatch.java
@@ -1,4 +1,0 @@
-package net.minestom.server.instance.batch.v2;
-
-public class RelativeBlockBatch {
-}
